# docs-site/openapi/mobile-lan.yaml
openapi: 3.1.0
info:
  title: Clipboard Mobile LAN API
  version: '1.0.0'
  summary: SyncClipboard-compatible LAN endpoints exposed by the desktop daemon.
  description: |
    The desktop daemon exposes a small SyncClipboard-compatible HTTP service on the
    LAN so a paired mobile client (iOS Shortcuts, Android, or any SyncClipboard
    client) can read and push clipboard items.

    Every endpoint requires **HTTP Basic Authentication** except `GET /healthz`.
    Credentials are minted on the desktop side when the user pairs a mobile device.
    Username and password are bound to a single registered device. Unauthenticated
    or wrong-credential requests get `401 Unauthorized` with
    `WWW-Authenticate: Basic`.

    `GET /healthz` is public by design: it is a constant-cost liveness probe for
    offline clients, and authenticating it would mean running an Argon2id
    verification per probe. It returns no identifying information — see its
    description below.

    The wire format follows the SyncClipboard project's JSON schema. Field casing
    is fixed by that protocol; the desktop also accepts PascalCase aliases on
    the request side because iOS Shortcuts emits `Type` (PascalCase) for the
    `type` field while the rest of the body is camelCase.

    Metadata-only requests such as `PUT /SyncClipboard.json` are capped at
    **16 MiB**. Binary uploads through `PUT /file/{dataName}` and multipart
...
    per uploaded file.
  contact:
    name: Clipboard
...
servers:
...
security:
  - basicAuth: []
paths:
...
  /SyncClipboard.json:
    get:
      tags: [Clipboard]
      operationId: getSyncClipboard
      summary: Fetch the latest clipboard item
      description: |
        Returns the most recent clipboard entry on the desktop, translated into
        the SyncClipboard wire DTO. If the desktop has no clipboard item yet,
        the response is an empty Text document for compatibility.

        For Image / File items, the binary payload is fetched separately via
        `GET /file/{dataName}`.
      responses:
        '200':
          description: Latest clipboard entry, or an empty Text document when no entry exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncClipboardDoc'
              examples:
                text:
                  summary: Text clipboard
                  value:
                    type: Text
                    text: hello from desktop
                    hasData: false
                    size: 18
                    hash: 2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824
                image:
                  summary: Image clipboard
                  value:
                    type: Image
                    text: ''
                    dataName: clipboard.png
                    hasData: true
                    size: 142536
                    hash: 9b2cf964fa8c0b5a...
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Clipboard snapshot port failure on the desktop.
    put:
      tags: [Clipboard]
      operationId: putSyncClipboard
      summary: Push a clipboard item from the mobile device
      description: |
        Accepts a SyncClipboard wire DTO. The body is decoded, mapped into the
        application model, and applied through the inbound apply pipeline as if
        it came from a peer device.

        Both camelCase and PascalCase keys are accepted (`Type`, `HasData`,
        `DataName`, etc.) because the iOS Shortcuts client emits a mixed-case
        body.

        Decode failure produces `400 Bad Request`; metadata payloads larger
        than 16 MiB produce `413 Payload Too Large`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncClipboardDoc'
            examples:
              text:
                summary: Push plain text
                value:
                  type: Text
                  text: hello from phone
                  hasData: false
                  size: 16
              file:
                summary: Push file metadata (PascalCase as emitted by iOS Shortcuts)
                value:
                  hasData: true
                  Type: File
                  dataName: foo.pdf
                  text: foo.pdf
      responses:
        '200':
          description: |
            Applied. The classic SyncClipboard protocol never defined a PUT
            response body (an empty `200` is the historical contract), so this
            is a purely additive extension: legacy and third-party
            SyncClipboard clients that only check the status code ignore it
            safely. Clipboard's own mobile sync engine reads `contentId`
            back to learn the server identity of the write it just made,
            without a follow-up `GET`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncClipboardPutAck'
              examples:
                applied:
                  summary: Applied with a learned content identity
                  value:
                    contentId: blake3v1:9b2cf964fa8c0b5a...
        '400':
          description: Body could not be parsed as a SyncClipboard document or `type` value is unknown.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Body exceeds 16 MiB.
        '500':
          description: Inbound apply pipeline failure on the desktop.
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Username and password are minted on the desktop during device pairing
        and are bound to a single mobile device record.
  responses:
    Unauthorized:
      description: |
        Missing, malformed, or wrong Basic credentials. The response carries
        `WWW-Authenticate: Basic`.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: 'Basic realm="Clipboard"'
  schemas:
    SyncClipboardDoc:
      type: object
      description: |
        SyncClipboard wire DTO. Field casing is fixed by the SyncClipboard
        project's protocol: the value of `type` is PascalCase, all keys are
        camelCase. On the request side the desktop also accepts PascalCase
        aliases (`Type`, `Text`, `DataName`, `HasData`, `Size`, `Hash`).
      required: [type, hasData, size]
      properties:
        type:
          description: Item kind. PascalCase by SyncClipboard convention.
          type: string
          enum: [Text, Image, File, Group]
        text:
          description: |
            For Text items, the clipboard text. For Image / File items, this
            field is typically the displayed filename or empty.
          type: string
          default: ''
        dataName:
          description: |
            Filename associated with the binary payload. Required for Image /
            File items, omitted for plain Text.
          type: string
          nullable: true
        hasData:
          description: |
            True iff a binary payload accompanies this entry (i.e. an Image or
            File item with bytes available via `/file/{dataName}`).
          type: boolean
          default: false
        size:
          description: Size in bytes of the binary payload (0 for plain Text).
          type: integer
          format: int64
          minimum: 0
          default: 0
        hash:
          description: |
            SHA-256 hex digest of the wire content. Mobile clients may omit it
            on `PUT`; the desktop fills it in on `GET` for SyncClipboard
            desktop-client compatibility.
          type: string
          nullable: true
          example: 2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824
    SyncClipboardPutAck:
      type: object
      description: |
        Response body for `PUT /SyncClipboard.json`. Purely additive relative
        to the classic SyncClipboard protocol (historically an empty `200`);
        legacy clients that don't parse the body are unaffected.
      properties:
        contentId:
          description: |
            Server-assigned stable identity for the content that was just
            written. Present when the write was applied or recognized as a
            duplicate of existing content; omitted if the daemon could not
            determine one for this response.
          type: string
          nullable: true
          example: blake3v1:9b2cf964fa8c0b5a...
...